Moved model development tools into their own Maven module.
This will be for things like fake data sources, visualization tools and
so forth.
