+++ pom.xml
@@ -32,6 +32,7 @@
     <modules>
         <module>core</module>
         <module>kafka</module>
+        <module>tools</module>
     </modules>
 
     <properties>
