+++ tools/pom.xml
@@ -27,17 +27,27 @@
 
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.expedia.adaptivealerting</groupId>
-    <artifactId>adaptive-alerting-core</artifactId>
+    <artifactId>adaptive-alerting-tools</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>${project.groupId}:${project.artifactId}</name>
-    <description>Anomaly detection for streaming time series.</description>
+    <description>Model development tools.</description>
     <url>https://github.com/ExpediaDotCom/haystack-adaptive-alerting</url>
 
     <dependencies>
 
         <!-- Haystack -->
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>adaptive-alerting-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>adaptive-alerting-kafka</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>com.expedia.www</groupId>
             <artifactId>haystack-commons</artifactId>
@@ -45,12 +55,20 @@
 
         <!-- Third-party -->
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
+            <groupId>org.apache.kafka</groupId>
+            <artifactId>kafka-streams</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.msgpack</groupId>
+            <artifactId>msgpack-core</artifactId>
         </dependency>
         <dependency>
             <groupId>com.opencsv</groupId>
             <artifactId>opencsv</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
