+++ tools/src/test/java/com/expedia/adaptivealerting/tools/metricsource/ConsoleLogMetricSubscriberTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expedia.adaptivealerting.core.metricsource;
+package com.expedia.adaptivealerting.tools.metricsource;
 
 import com.expedia.www.haystack.commons.entities.MetricPoint;
 import com.expedia.www.haystack.commons.entities.MetricType;
